code/requirements.txt
boto3
mysql.connector